connect freamwork Hibernet and create db

+++ src/test/java/io/github/mat3e/HelloServiceTest.java
@@ -43,7 +43,7 @@ public class HelloServiceTest {
         // given
         var mockRepository = new LangRepository() {
             @Override
-            Optional<Lang> findById(Long id) {
+            Optional<Lang> findById(Integer id) {
                 return Optional.empty();
             }
         };
@@ -89,7 +89,7 @@ public class HelloServiceTest {
     private LangRepository fallbackLangIdRepository (){
         return new LangRepository() {
             @Override
-            Optional<Lang> findById(Long id) {
+            Optional<Lang> findById(Integer id) {
                 if (id.equals(HelloService.FALLBACK_LANG.getId())){
                     return Optional.of(new Lang(null, FALLBACK_ID_WELCOME, null));
                 }
@@ -101,7 +101,7 @@ public class HelloServiceTest {
     private LangRepository alwaysReturingHelloRepository() {
         return new LangRepository() {
             @Override
-            Optional<Lang> findById(Long id) {
+            Optional<Lang> findById(Integer id) {
                 return Optional.of(new Lang(null, WELCOME, null));
             }
         };
